refactor(desktop): the selected agent home moves beside its peers
A second consumer is about to read it, so a hook that only the workspace
component could reach becomes a domain hook under features/session/hooks.
